requirements.txt
fastapi==0.115.14
uvicorn[standard]
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic==2.11.7
scikit-learn
pandas
numpy